Add pubkey service
