+++ views/src/home/home.html
@@ -96,9 +96,7 @@
                 </div>
 
                 <div class="[ text-center ]" style="margin: 15px;">
-                  <hearts [infoloaded]="heartemitter"
-                          [pubkeys]="pubkeys">
-                  </hearts>
+                  <hearts> </hearts>
                 </div>
 
             </div>
